requirements.txt
python-telegram-bot==20.2
python-dotenv~=1.0.0
replicate
